Rework std::format detection
Instead of including <format> and maybe running in an error or a message
that it is not supported (e.g. in pre C++20 mode), include <version> and
check the macro from there.

That way we declare std::format absent if there is no <verion> header,
but I think that's fine.